livecheck: rename version.rb to livecheck_version.rb
