+++ Library/Homebrew/livecheck/livecheck_version.rb
